Vue3: Migrate event handling
Vue 3 doesn't support the $listeners anymore. Events must be declared in
advance, and event handlers are passed as props, with the 'on' prefix.
Instead of parsing props names, and conditionally bind map events, we're
just binding all map events. This is the Vue way. Components just emit
events whether someone is listening or not.

Add `emits` options to declare all emitted events.